+++ rollup.config.js
@@ -1,0 +1,36 @@
+import peerDepsExternal from 'rollup-plugin-peer-deps-external'
+import resolve from "@rollup/plugin-node-resolve";
+import typescript from "rollup-plugin-typescript2";
+import commonjs from '@rollup/plugin-commonjs';
+import postcss from "rollup-plugin-postcss";
+// import { uglify } from "rollup-plugin-uglify";
+import { babel } from '@rollup/plugin-babel';
+
+export default {
+    input: "./src/index.tsx",
+    output: [
+        {
+            file: 'dist/index.js',
+            format: 'cjs',
+            sourcemap: true
+        },
+        {
+            file: 'dist/index.esm.js',
+            format: 'esm',
+            sourcemap: true
+        }
+    ],
+    plugins: [
+        peerDepsExternal(),
+        postcss({
+            inject: true
+        }),
+        typescript(),
+        babel({
+            extensions: ['.js', '.jsx', '.ts', '.tsx'],
+            exclude: 'node_modules/**'
+        }),
+        resolve(),
+        commonjs()
+    ],
+}