+++ scripts/run_app_frontend.sh
@@ -42,9 +42,11 @@ if [ "$1" = "dev" ]; then
 fi
 if [ "$1" = "dev" ]; then
     if [ "${RUN_METHOD}" = "webpack" ]; then
+        perl -i -pe"s|\"type\": \"module\"|\"type1\": \"module\"|g" package.json
         perl -i -pe"s|\%PUBLIC_URL\%||g" public/index.html
         npm run start-dev-webpack
         cp public/index-template.html public/index.html
+        perl -i -pe"s|\"type1\": \"module\"|\"type\": \"module\"|g" package.json
     else
         export REACT_APP_VERSION=$(cat version.txt)
         echo "REACT_APP_VERSION = ${REACT_APP_VERSION}"